requirements.txt
SQLAlchemy==1.4.31
discord.py==1.7.3
dataclass_factory==2.14
PyYAML==6.0
psycopg2-binary==2.9.3
